A LOT OF CHANGES: Game now owns a level. And initialises the Game to Level1
Declarations: Holds a map for the items and a map for the loaded images. Is now persistent and can be accessed from Level

Level: Level has been refactored to breakdown the handling of Reading the Xml and Constructing the items to push back into Game->mItems.

Digit, Given, Container, Xray, and Sparty have their constructors refactored to allow initialisation of their attributes

Check the attributes of Digit, Given, Container, Xray, and Sparty to know what's available to you now. I will create the tests some other day but this should be the new approach to Items. Hopefully we can reevaluate old code.

I would say the code is easier to read now but feel free to ask questions.
Please let me know if your game is lagy